# apps/wiki/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plunk API",
    "description": "Open-source email platform API for transactional emails, campaigns, and marketing automation",
    "version": "1.0.0",
    "contact": {
      "name": "Plunk Support",
      "url": "https://useplunk.com"
    }
  },
  "servers": [
    {
      "url": "https://api.useplunk.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key authentication. **Secret keys (sk_*)** are required for all endpoints except `/v1/track`. **Public keys (pk_*)** only work with the `/v1/track` endpoint for client-side event tracking. The project is automatically derived from the key."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "time": {
            "type": "integer"
          }
        }
      }
    },
...
  },
  "paths": {
    "/v1/send": {
      "post": {
        "tags": ["Public API"],
        "summary": "Send transactional email",
        "description": "Send a transactional email via the public API. Automatically creates/updates contacts.",
        "operationId": "sendEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["to", "subscribed"],
                "properties": {
                  "to": {
                    "oneOf": [
                      {
                        "type": "string",
                        "format": "email"
                      },
                      {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "format": "email"
                        }
                      }
                    ],
                    "description": "Recipient email(s)"
                  },
                  "subject": {
                    "type": "string",
                    "description": "Email subject (required if no template)"
                  },
                  "body": {
                    "type": "string",
                    "description": "Email body HTML (required if no template)"
                  },
                  "template": {
                    "type": "string",
                    "description": "Template identifier to use"
                  },
                  "from": {
                    "type": "string",
                    "format": "email",
                    "description": "Custom from address (requires verified domain)"
                  },
                  "subscribed": {
                    "type": "boolean",
                    "description": "Whether recipient is subscribed"
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Template variables"
                  },
                  "headers": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "Custom email headers"
                  },
                  "reply": {
                    "type": "string",
                    "format": "email",
                    "description": "Reply-to address"
                  }
                }
              },
              "examples": {
                "simple": {
                  "summary": "Simple email",
                  "value": {
                    "to": "user@example.com",
                    "subject": "Welcome!",
                    "body": "<h1>Hello {{name}}</h1>",
                    "subscribed": true,
                    "data": {
                      "name": "John Doe"
                    }
                  }
                },
                "withTemplate": {
                  "summary": "Using template",
                  "value": {
                    "to": "user@example.com",
                    "template": "welcome-email",
                    "subscribed": true,
                    "data": {
                      "firstName": "John",
                      "lastName": "Doe"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email queued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "emails": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "to": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/track": {
      "post": {
        "tags": ["Public API"],
        "summary": "Track custom event",
        "description": "Track a custom event for a contact. Automatically creates/updates the contact.",
        "operationId": "trackEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "event"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Contact email"
                  },
                  "event": {
                    "type": "string",
                    "description": "Event name"
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Event data"
                  }
                }
              },
              "example": {
                "email": "user@example.com",
                "event": "purchase",
                "data": {
                  "product": "Premium Plan",
                  "amount": 99.00
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event tracked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "contact": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            }
                          }
                        },
                        "event": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "event": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}